+++ tests/test_project.py
@@ -13,7 +13,7 @@ def test_project():
     assert project.path().endswith("samples/project")
     project.__enter__()
     assert Project.get().name() == "project"
-    assert Project.get().module == "_machinable.project"
+    assert Project.get().module == "interface.project"
     project.__exit__()
     # note that this may fail if other tests have errors
     # and failed to clean up the project
@@ -37,12 +37,10 @@ def test_project_events(tmp_storage):
     component.launch()
 
     # remotes
-    shutil.rmtree(
-        "tests/samples/project/_machinable/remotes", ignore_errors=True
-    )
+    shutil.rmtree("tests/samples/project/interface/remotes", ignore_errors=True)
     Component.instance("!hello")()
     Component.instance("!hello-link")()
-    assert os.path.exists("tests/samples/project/_machinable/remotes")
+    assert os.path.exists("tests/samples/project/interface/remotes")
     with pytest.raises(ValueError):
         Component.instance("!invalid")